# sutra/contracts/sutra.openapi.yaml
openapi: 3.0.3
info:
  title: sutra
  description: The single API surface — CLI, web UI, and external agents all speak this contract.
  version: 0.2.0
paths:
...
  /reviews:
    post:
      operationId: createReview
      summary: Submit a review
      description: >-
        Names an issue and exactly one deliverable, branch or doc version,
        optionally stamped with a session id; created in state open and visible
        to reviewers (AC-review-create, AC-review-deliverable-kinds). For a
        code deliverable, `base_commit` — the merge base with the project's
        default branch — is resolved and pinned at submission time, not
        client-supplied, so the rendered diff never shifts as the default
        branch moves. A code deliverable with an unset project `repo_path`
        or an unresolvable `commit`/`base_commit` is rejected 409 at
        submission time — as is one whose pinned diff cannot be rendered
        within the server's 10 MiB deliverable limit, so every accepted
        review's content is openable by a reviewer. A supplied `session`
        is recorded on the submission.
      tags: [reviews]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewReview'
      responses:
        '201':
          description: Review created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      operationId: listReviews
      summary: List reviews
      description: >-
        Filtering by session returns the reviews stamped with that instance
        (AC-search-session) — a session query matches a review whose ANY
        ReviewSubmission.session equals the query, not just the latest
        submission's session, since the session that mattered may belong to
        an earlier revision.
      tags: [reviews]
      parameters:
        - name: issue
          in: query
          schema: { type: string, format: uuid }
        - name: session
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { $ref: '#/components/schemas/ReviewState' }
      responses:
        '200':
          description: Reviews.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Review' }
  /reviews/{reviewId}:
    get:
      operationId: getReview
      summary: Get a review
      description: Shows the deliverable's content for reading (AC-review-web).
      tags: [reviews]
      parameters:
        - $ref: '#/components/parameters/ReviewIdPath'
      responses:
        '200':
          description: The review.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          $ref: '#/components/responses/NotFound'
  /reviews/{reviewId}/verdict:
    post:
      operationId: setReviewVerdict
      summary: Set a review's verdict
      description: >-
        A human sets approved or changes-requested; the state changes and an
        Event records who and when (AC-review-verdict). The verdict applies
        to the review's current revision — for a code deliverable, approval
        covers the commit pinned as of that revision. A verdict carrying the
        current revision always applies, including changing an earlier
        verdict (approved -> changes-requested and vice versa); the latest
        verdict for the current revision is authoritative, and each verdict
        is recorded as an event — until the approval is consumed
        (AC-review-consume): a verdict on a consumed review is rejected with
        a conflict and no mutation. Approval publishes an event a subscriber
        can act on (AC-review-approved-event);
        changes-requested carries the review's session id and issue ref so a
        resubmission routes back to the same agent instance
        (AC-review-rework). Invoked by the review view's verdict action
        (ACT-verdict).
      tags: [reviews]
      parameters:
        - $ref: '#/components/parameters/ReviewIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewVerdict'
      responses:
        '200':
          description: Review with the new state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /reviews/{reviewId}/consume:
    post:
      operationId: consumeReviewApproval
      summary: Atomically consume an approved review's approval
      description: >-
        Conditional on state approved at the expected current revision;
        stamps a durable consumption that permanently fences verdict changes
        and atomically records a `review.consumed` event carrying the actor,
        review, and consumed revision — the audit trail shows who froze the
        verdict and when (AC-review-consume). Idempotent ONLY under the original
        Idempotency-Key — a replay under that key returns the original
        success; a DISTINCT consumption attempt (different key, any actor)
        against an already-consumed approval is rejected with a conflict,
        so multiple subscribers can never each receive success and act on
        the same approval. Fails with a
        conflict — consuming nothing — if the verdict changed or the
        revision moved first. Subscribers call this before acting on an
        approval (e.g. merging), so approval cannot be reversed mid-action.
      tags: [reviews]
      parameters:
        - $ref: '#/components/parameters/ReviewIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expected_revision, expected_verdict_event, actor]
              properties:
                expected_revision: { type: integer }
                expected_verdict_event:
                  type: string
                  format: uuid
                  description: The approval's verdict event; rejected with a conflict if it is no longer the review's latest — a verdict ABA invalidates delayed consumptions even at a matching revision and state.
                actor: { type: string, format: uuid }
      responses:
        '200':
          description: Review with the consumption stamped.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /reviews/{reviewId}/resubmit:
    post:
      operationId: resubmitReview
      summary: Resubmit a review with a replacement deliverable
      description: >-
        Valid only from state changes-requested; resubmitting an open or
        approved review is rejected with a conflict, as is an
        expected_revision that does not match the review's current revision
        — atomically, so a delayed resubmission that never originally landed
        cannot apply stale content after the review advanced. After
        changes-requested, the agent resubmits a replacement deliverable —
        the same exactly-one shape as creation, a branch pinned at an
        immutable commit or a doc version — and the review returns to state
        open, incrementing its revision; feedback left on the prior
        deliverable stays associated with the revision it concerned
        (AC-review-rework). Appends a new ReviewSubmission recording what
        this revision submitted; prior submissions are preserved
        immutably. A supplied `session` is recorded on the submission. For
        a code deliverable, `base_commit` is resolved and pinned at
        submission time, not client-supplied. A code deliverable with an
        unset project `repo_path` or an unresolvable `commit`/`base_commit`
        is rejected 409 at submission time.
      tags: [reviews]
      parameters:
        - $ref: '#/components/parameters/ReviewIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResubmitReview'
      responses:
        '200':
          description: Review reopened with the replacement deliverable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  parameters:
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Client-supplied key that makes this mutating operation safe to retry
        (CON-idempotent-mutations): replaying the same key returns the
        original response verbatim and has no second effect — no duplicate
        resource, no second state transition, no second side effect.
      schema: { type: string }
...
    ReviewIdPath:
      name: reviewId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: >-
        The request is invalid — e.g. an actor or assignee identity id that
        names no known identity (AC-identity-referenced), or a status outside
        the five defined values (AC-status-invalid). The API deals only in
        identity ids; resolving a human-typed handle to an id is the CLI's job.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The referenced resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: >-
        The request conflicts with current state — e.g. closing without an
        approved review, a blocking or ancestry cycle, a colliding import, a
        write against an archived project, or a review-anchored comment or
        verdict naming a stale revision (superseded by a resubmission since
        the reviewer last saw it) — rejected atomically, never applied to
        content the reviewer did not see.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    ReviewState:
      type: string
      enum: [open, changes-requested, approved]
...
    NewReview:
      type: object
      description: >-
        Exactly one deliverable: a branch pinned at an immutable commit
        (both required together; branch is display metadata, commit is what
        approval covers), or a doc version.
      required: [issue, author]
      properties:
        issue: { type: string, format: uuid }
        expected_base_commit:
          type: string
          description: >-
            Code deliverables only — the merge base the caller expects.
            Validated against the base_commit resolved at submission time;
            a mismatch rejects 409 and no review, submission, or event
            exists. The fence catches STALE CALLER KNOWLEDGE — the
            repository is an external store no tracker transaction can
            lock, so validation is against the ref as observed at
            submission (bounded staleness), and the pinned base_commit
            remains authoritative for what approval covers regardless of
            concurrent branch movement.
        expected_default_head:
          type: string
          description: >-
            Code deliverables only — the default-branch head the caller
            expects. Validated against the ref as observed at submission
            time; a mismatch rejects 409 with nothing created. The head,
            not the merge base, is the real movement fence — a
            fast-forwarded default branch keeps the old merge base. As
            with expected_base_commit, the repository is external: the
            fence rejects stale caller observations; it does not — and
            cannot — serialize against pushes concurrent with the
            submission transaction.
        branch: { type: string }
        commit:
          type: string
          pattern: "^[0-9a-f]{40}([0-9a-f]{24})?$"
          description: >-
            Canonical full object id — never an abbreviation. Immutable,
            pinned at creation; required alongside branch.
        doc_version: { type: string, format: uuid }
        session: { type: string }
        summary: { type: string }
        author: { type: string, format: uuid }
      oneOf:
        - required: [branch, commit]
          not:
            required: [doc_version]
        - required: [doc_version]
          not:
            anyOf:
              - required: [branch]
              - required: [commit]
...